requirements.txt
llama-index
llama-index-core
llama-index-llms-gemini
llama-index-embeddings-gemini
langchain-core
langchain-google-genai
langgraph
langgraph-checkpoint-sqlite
python-dotenv
fastapi
uvicorn
pyodbc
docx2txt
llama-index-readers-azstorage-blob
python-multipart
python-docx
